Change Token balance to be of type BigInteger

+++ src/main/java/com/enjin/enjincoin/sdk/model/service/tokens/Token.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.ToString;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @ToString
 @Getter
@@ -49,7 +50,7 @@ public class Token {
 
     private Boolean isCreator;
 
-    private BigDecimal balance;
+    private BigInteger balance;
 
     private String index;
